feat: add GitHub Actions workflows for livetrace and startled CLIs (#122)
* feat: add GitHub Actions workflows for livetrace and startled CLIs

- Introduced `pr-check-livetrace.yml` for pull request checks on the livetrace CLI, including quality checks and build attempts.
- Added `pr-check-startled.yml` for pull request checks on the startled CLI, with similar quality checks and build attempts.
- Created `release-livetrace.yml` for automated releases of the livetrace CLI based on version tags.
- Created `release-startled.yml` for automated releases of the startled CLI based on version tags.
- Added `reusable-release-cli.yml` to define a reusable workflow for building and releasing CLI applications.

These workflows enhance CI/CD processes for both CLIs, ensuring code quality and streamlined releases.

* fix: update GitHub Actions workflows to remove version pinning for reusable workflows

* chore: update dependencies in Cargo.toml and modify rotel layer version in template.yaml

- Added `pulldown-cmark` version 0.13.0 and `rust_decimal` version 1.37.1 to Cargo.toml.
- Updated the rotel layer version in template.yaml from 20 to 21.
- Added .gitignore file to confmaps directory to exclude files containing 'secret'.

* refactor: improve formatting in benchmark results and report template checks

- Adjusted formatting in `console.rs` for better readability of benchmark results, specifically aligning standard deviation values.
- Simplified checks for the existence of essential HTML templates in `report.rs` by using `any()` instead of `find()`, enhancing code clarity.

* chore: update Rust version in GitHub Actions workflow

* chore: update GitHub Actions workflow and fix memory size handling in lambda

- Changed the OS names in the GitHub Actions workflow from 'ubuntu-latest' to 'ubuntu-24.04' and 'ubuntu-latest-arm64' to 'ubuntu-24.04-arm' for better version specificity.
- Removed unnecessary type casting of memory size in `lambda.rs`, simplifying the code.

* fix: correct typo in OS name in GitHub Actions workflow

+++ cli/startled/src/lambda.rs
@@ -358,7 +358,7 @@ pub async fn get_function_config(
         )
     })?;
     Ok(OriginalConfig {
-        memory_size: config.memory_size().unwrap_or(128) as i32,
+        memory_size: config.memory_size().unwrap_or(128),
         environment: config
             .environment()
             .and_then(|e| e.variables())
